Changed LSAPLookupLevel return value to short and added unit tests

+++ src/main/java/com/rapid7/client/dcerpc/mslsad/LSAPLookupLevel.java
@@ -242,7 +242,7 @@ public enum LSAPLookupLevel {
      */
     LsapLookupRODCReferralToFullDC;
 
-    public int getValue() {
-        return ordinal() + 1;
+    public short getValue() {
+        return (short) (ordinal() + 1);
     }
 }
